+++ lorien/Misc/Config.gd
@@ -9,6 +9,7 @@ const CONFIG_PATH 					:= "user://settings.cfg"
 const DEFAULT_CANVAS_COLOR 			:= Color("202124")
 const DEFAULT_BRUSH_COLOR 			:= Color.white
 const DEFAULT_BRUSH_SIZE			:= 12
+const DEFAULT_PRESSURE_SENSITIVITY  := 1.0
 const DEFAULT_AA_MODE				:= Types.AAMode.TEXTURE_FILL
 const DEFAULT_SELECTION_COLOR 		:= Color("#2a967c")
 const DEFAULT_FOREGROUND_FPS 		:= 144